pyproject.toml
[project]
name = "airflow-lms"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "apache-airflow-providers-mongo>=5.3.2",
    "apache-airflow-task-sdk>=1.1.7",
    "pandas>=3.0.1",
]
